fix(#363): address PR review — common embeddings, persist cap, investigate bug
Move shared embedding utilities to common/, enforce assignment cap in persist noise fallback, expose row_to_features, fix investigate_363 section 15 loop.
